pyproject.toml
[project]
name = "guardrails"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "windows-curses>=2.4.1",
]
